+++ test-roast.js
@@ -1,6 +1,6 @@
 "use strict";
 /**
- * Test file for VS Roast Extension
+ * Test file for Roast Extension
  * This file contains various code patterns that should trigger roasts
  */
 Object.defineProperty(exports, "__esModule", { value: true });